requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.27
gunicorn==23.0.0
mysql-connector-python==9.0.0
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.5.5
typing-extensions==4.13.2
tzdata==2025.3
whitenoise==6.7.0
